# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NimbleWallet API",
    "description": "Wallet backend (Spring Modulith): authentication, ledger,\ndeposits (Razorpay), peer transfers, and activity/analytics.\nSession is a JWT carried in the httpOnly `Authentication` cookie.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/webhooks/razorpay": {
      "post": {
        "tags": [
          "razorpay-webhook-controller"
        ],
        "operationId": "capture",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RazorpayCapturePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transfer/p2p": {
      "post": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "transfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/deposit/order": {
      "post": {
        "tags": [
          "deposit-controller"
        ],
        "operationId": "createOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DepositOrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/deposit/fail": {
      "post": {
        "tags": [
          "deposit-controller"
        ],
        "operationId": "fail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/sign-up": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "signUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/me": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/deposit/history": {
      "get": {
        "tags": [
          "deposit-controller"
        ],
        "operationId": "history",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepositHistoryEntry"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/balance": {
      "get": {
        "tags": [
          "balance-controller"
        ],
        "operationId": "balance",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "googleLogin",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/activity": {
      "get": {
        "tags": [
          "activity-controller"
        ],
        "operationId": "statement",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StatementView"
                }
              }
            }
          }
        }
      }
    },
    "/activity/p2p-recipients": {
      "get": {
        "tags": [
          "activity-controller"
        ],
        "operationId": "recentRecipients",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipientView"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activity/p2p-count": {
      "get": {
        "tags": [
          "activity-controller"
        ],
        "operationId": "peerTransferCount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/activity/analytics": {
      "get": {
        "tags": [
          "activity-controller"
        ],
        "operationId": "analytics",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsView"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RazorpayCapturePayload": {
        "type": "object",
        "properties": {
          "razorpay_order_id": {
            "type": "string"
          },
          "razorpay_payment_id": {
            "type": "string"
          },
          "razorpay_signature": {
            "type": "string"
          }
        }
      },
      "TransferRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CreateOrderRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DepositOrderResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "currency": {
            "type": "string"
          },
          "keyId": {
            "type": "string"
          }
        }
      },
      "FailRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "SignUpRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8
          },
          "number": {
            "type": "string",
            "pattern": "\\d{10}"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserView"
          }
        }
      },
      "UserView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "number": {
            "type": "string"
          }
        }
      },
      "DepositHistoryEntry": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "completedCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalAmountMinor": {
            "type": "integer",
            "format": "int64"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BalanceSnapshot": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "locked": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DepositEntry": {
        "type": "object",
        "properties": {
          "amountMinor": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "orderToken": {
            "type": "string"
          }
        }
      },
      "StatementView": {
        "type": "object",
        "properties": {
          "transfers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferEntry"
            }
          },
          "deposits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DepositEntry"
            }
          }
        }
      },
      "TransferEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "amountMinor": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "direction": {
            "type": "string"
          },
          "counterpartyName": {
            "type": "string"
          },
          "counterpartyEmail": {
            "type": "string"
          }
        }
      },
      "RecipientView": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          }
        }
      },
      "AnalyticsView": {
        "type": "object",
        "properties": {
          "dailyTotals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyTotal"
            }
          },
          "totalInMinor": {
            "type": "integer",
            "format": "int64"
          },
          "totalOutMinor": {
            "type": "integer",
            "format": "int64"
          },
          "netMinor": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DailyTotal": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "inflowMinor": {
            "type": "integer",
            "format": "int64"
          },
          "outflowMinor": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "name": "Authentication",
        "in": "cookie"
      }
    }
  }
}